refactor: replace placeholder layout components (#14)
* refactor: replace placeholder components

Replaces placeholder components from the starter kit with actual ones
from chakra-ui.

* refactor: more refactoring
