--- components/PageBody.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react'
-
-import styles from './PageBody.module.css'
-
-const PageBody: React.FC = ({ children }) => {
-  return <section className={styles.root}>{children}</section>
-}
-
-export default PageBody
