--- components/PageHeader.module.css
@@ -1,7 +1,0 @@
-.root {
-  padding: 2rem 1.5rem 1rem;
-}
-
-.title {
-  margin: 0;
-}
