+++ pages/index.tsx
@@ -2,9 +2,7 @@ import * as React from 'react'
 import { NextPage } from 'next'
 import Link from 'next/link'
 
-import Page from '~/components/Page'
-import PageHeader from '~/components/PageHeader'
-import PageBody from '~/components/PageBody'
+import { Page, PageHeader, PageBody } from '~/components/Layout'
 import Navigation from '~/components/Navigation'
 import Content from '~/components/Content'
 import relativeTime from '~/utils/relativeTime'
@@ -15,7 +13,7 @@ const IndexPage: NextPage = () => (
   <Page>
     <Navigation />
     <Content>
-      <PageHeader>Hello world.</PageHeader>
+      <PageHeader title="Hello world." />
       <PageBody>
         <p>
           Lorem ipsum dolor sit amet, consectetur adipiscing elit. Certe non potest. Negabat igitur ullam esse artem, quae ipsa a se
